refactor(website): decouple tickets-db from website-db (G-CQ07 cycle #1) [T001172]
